support larger numbers for time

+++ include/engine.hpp
@@ -13,14 +13,14 @@ public:
     struct GoOptions {
         std::optional<std::vector<Move>> searchmoves{std::nullopt};
         bool ponder{false};
-        std::optional<int> wtime{std::nullopt};
-        std::optional<int> btime{std::nullopt};
-        std::optional<int> winc{std::nullopt};
-        std::optional<int> binc{std::nullopt};
+        std::optional<int64_t> wtime{std::nullopt};
+        std::optional<int64_t> btime{std::nullopt};
+        std::optional<int64_t> winc{std::nullopt};
+        std::optional<int64_t> binc{std::nullopt};
         std::optional<int> movestogo{std::nullopt};
         std::optional<int> depth{std::nullopt};
-        std::optional<int> nodes{std::nullopt};
-        std::optional<int> movetime{std::nullopt};
+        std::optional<int64_t> nodes{std::nullopt};
+        std::optional<int64_t> movetime{std::nullopt};
         bool infinite{false};
     };
